services/product_service/requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Werkzeug==3.1
python-consul